+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [2.3] (01-Jan-2025)
 
 ### Changed
 - MBean name of ClasspathAgent changed to "clazzfish:type=agent,agent=ClasspathAgent"